refactor: trim dead weight — decouple runtime-portability, drop self-imports
Two audit-flagged code-hygiene cleanups (deferred items).

1. cloud-server runtime-portability.ts (997 lines) exported the entire
   sandbox-runtime surface, but the production checkpoint path needs only 2
   symbols. Extract the snapshot-portability cluster
   (isRuntimeSnapshotSecretBearingPath, PortableRuntimeEntry* types,
   buildPortableRuntimeManifest, runtimePathsForPortability) into a new
   runtime-snapshot-portability.ts (149 lines) and point workspace-checkpoint-store
   at it, so the production path no longer loads the ~860-line sandbox module.
   runtime-portability.ts re-exports the moved symbols so scripts/tests are
   untouched. runtime-portability.ts: 997 -> 860 lines.

2. runtime-host imported from its own package name (@open-cowork/runtime-host
   and /config, 45 sites) instead of relative paths — a self-barrel smell.
   Converted all to direct relative imports of the defining module. One import
   is deliberately kept: runtime-managed-server-supervisor.ts is forked as raw
   TS in tests, where a relative `.js` specifier does not resolve under
   --experimental-strip-types; the package-name import resolves in both dist
   and strip-types contexts (documented inline).

No logic changes. Gates: typecheck, eslint --max-warnings 0, knip, node (2238)
+ workspace (313) suites, coverage inventory 92.0% all pass.
