+++ packages/cloud-server/src/workspace-checkpoint-store.ts
@@ -1,8 +1,8 @@
 import { createHash, randomUUID } from 'node:crypto'
 import { readdir, readFile, lstat, mkdir, realpath, writeFile } from 'node:fs/promises'
 import { basename, dirname, isAbsolute, join, relative, resolve } from 'node:path'
-import type { PortableRuntimeEntryKind } from './runtime-portability.ts'
-import { isRuntimeSnapshotSecretBearingPath } from './runtime-portability.ts'
+import type { PortableRuntimeEntryKind } from './runtime-snapshot-portability.ts'
+import { isRuntimeSnapshotSecretBearingPath } from './runtime-snapshot-portability.ts'
 import type { ObjectStoreAdapter } from './object-store.ts'
 import { assertSafeObjectKey } from './object-store.ts'
 import type { PathProvider } from './path-provider.ts'
