+++ packages/runtime-host/src/runtime-mcp.ts
@@ -3,7 +3,7 @@ import { getAppPathHost } from '@open-cowork/shared/node'
 import { isAbsolute, join, resolve } from 'node:path'
 import { credentialFieldIsVisible } from '@open-cowork/shared'
 import type { CustomMcpConfig } from '@open-cowork/shared'
-import { getConfiguredMcpsFromConfig, type BundleMcp } from '@open-cowork/runtime-host/config'
+import { getConfiguredMcpsFromConfig, type BundleMcp } from './config-loader-core.js'
 import { getIntegrationCredentialValue, getEffectiveSettings, type CoworkSettings } from './settings.js'
 import { getMachineSkillsDir } from './runtime-paths.js'
 import { getAdcPathIfAvailable, getCachedAccessToken } from './auth.js'
