+++ packages/runtime-host/src/runtime-skill-catalog.ts
@@ -2,7 +2,7 @@ import { writeFileAtomic } from '@open-cowork/shared/node'
 import { mkdirSync, rmSync } from 'node:fs'
 import { dirname, join, resolve } from 'node:path'
 import type { RuntimeContextOptions } from '@open-cowork/shared'
-import { getProjectOverlayDirName } from '@open-cowork/runtime-host/config'
+import { getProjectOverlayDirName } from './config-loader-core.js'
 import { listEffectiveBuiltInSkillBundlesSync } from './effective-skills.js'
 import { warmBundledSkillIndex } from './bundled-skill-index.js'
 import { log } from '@open-cowork/shared/node'
