+++ packages/runtime-host/src/session-history-loader.ts
@@ -1,9 +1,9 @@
-import { normalizeSessionInfo, normalizeSessionMessages, normalizeSessionStatuses, type NormalizedSessionMessage } from '@open-cowork/runtime-host'
+import { normalizeSessionInfo, normalizeSessionMessages, normalizeSessionStatuses, type NormalizedSessionMessage } from './opencode-adapter.js'
 import { shortSessionId, asRecord, readRecordArray, readRecordValue, readString } from '@open-cowork/shared'
 import type { OpencodeClient, PermissionRequest, QuestionRequest } from '@opencode-ai/sdk/v2'
 import type { PendingApproval, PendingQuestion, SessionView } from '@open-cowork/shared'
 import { getClientForDirectory, getRuntimeHomeDir, getV2ClientForDirectory } from './runtime.js'
-import { getBrandName } from '@open-cowork/runtime-host/config'
+import { getBrandName } from './config-loader-core.js'
 import { getEffectiveSettings, loadSettings } from './settings.js'
 import { isInternalCoworkMessage } from './internal-message-utils.js'
 import { projectSessionHistory } from './session-history-projector.js'
