+++ packages/runtime-host/src/session-registry.ts
@@ -2,7 +2,7 @@ import { writeFileAtomic } from '@open-cowork/shared/node'
 import { existsSync, mkdirSync, readdirSync, readFileSync } from 'node:fs'
 import { join, resolve } from 'node:path'
 import type { SessionChangeSummary, SessionUsageSummary } from '@open-cowork/shared'
-import { getAppDataDir, getBrandName } from '@open-cowork/runtime-host/config'
+import { getAppDataDir, getBrandName } from './config-loader-core.js'
 import { log } from '@open-cowork/shared/node'
 import { getRuntimeHomeDir } from './runtime.js'
 import { isSandboxWorkspaceDir } from './runtime-paths.js'
